Made the beam_attr df type float so can get stats

+++ src/data/adcp_mission_analysis.py
@@ -288,7 +288,7 @@ def add_dive_averages(mission_summary, profiles_dict, combine=False):
     # hotfix as dive 5 appears to be a bench test...
     bar = adcp_df[adcp_df.cast_num != '0005a']
     baz = bar[bar.cast_num != '0005b']
-
+    beam_attrs = beam_attrs.astype(float)
     if combine:
         mission_summary = mission_summary.join(beam_attrs)
         return mission_summary, baz
